Partial windows support (will be unnecessary with the new linux subsystem)

+++ inc/classes/Archivers.php
@@ -34,15 +34,14 @@ class Archivers {
         $processes = explode("\n", shell_exec("tasklist.exe"));
         foreach ($processes as $process) {
           if (preg_match('/^(.*)\s+' . $pid . '/', $process)) {
-            $running = true;
-            break;
+            if (file_exists(Site::getPath() . "/backend/$board.kill")) {
+              return self::STOPPING;
+            } else {
+              return self::RUNNING;
+            }
           }
         }
-        if ($running == false) {
-          return self::STOPPED_UNCLEAN;
-        } else {
-          return self::RUNNING;
-        }
+        return self::STOPPED_UNCLEAN;
       }
     }
     return self::STOPPED;
@@ -57,12 +56,16 @@ class Archivers {
   static function getOutput(string $board):string {
     $status = self::getStatus($board);
     if($status == "Running" || $status == "Stopping") {
-      $path = Site::getPath()."/backend/$board.buff";
-      exec("screen -x $board -p0 -X hardcopy -h $path");
-      sleep(1);
-      $str = file_get_contents($path);
-      unlink($path);
-      return $str;
+      if (PHP_OS == "Linux") {
+        $path = Site::getPath()."/backend/$board.buff";
+        exec("screen -x $board -p0 -X hardcopy -h $path");
+        sleep(1);
+        $str = file_get_contents($path);
+        unlink($path);
+        return $str;
+      } else {
+        return file_get_contents(Site::getPath()."/backend/$board.log");
+      }
     }
     return "";
   }
@@ -102,16 +105,29 @@ class Archivers {
   }
 
   static function run(string $board):bool {
-    if (file_exists(Site::getPath() . "/backend/$board-archiver.php")) {
+    if (file_exists(Site::getPath() . "/backend/$board-archiver.php") 
+            && !file_exists(Site::getPath() . "/backend/no_custom_archivers")) {
       if (self::getStatus($board) == self::STOPPED || self::getStatus($board) == self::STOPPED_UNCLEAN) {
-        exec("cd " . Site::getPath() . "/backend/ && screen -dmS $board php $board-archiver.php");
+        if (PHP_OS == "Linux") {
+          exec("cd " . Site::getPath() . "/backend/ && screen -dmS $board php $board-archiver.php");
+        } else {
+          $path = Site::getPath() . "/backend/$board-archiver.php";
+          $cmd = "c:/php/php.exe \"$path\" -f";
+          pclose(popen('cd '.Site::getPath() . "/backend/".' && start /b '.$cmd,'r'));
+        }
         return true;
       }
     } else {
       Model::get()->getBoard($board);
       if (self::getStatus($board) == self::STOPPED || self::getStatus($board) == self::STOPPED_UNCLEAN) {
-        exec("cd " . Site::getPath() . "/backend/ && " .
-                "screen -dmS $board php archiver.php -b $board");
+        if (PHP_OS == "Linux") {
+          exec("cd " . Site::getPath() . "/backend/ && " .
+                  "screen -dmS $board php archiver.php -b $board");
+        } else {
+          $path = Site::getPath() . "/backend/archiver.php";
+          $cmd = "c:/php/php.exe \"$path\" -b $board -f";
+          pclose(popen('cd '.Site::getPath() . "/backend/".' && start /b '.$cmd,'r'));
+        }
         return true;
       }
     }
